+++ packages/access-api/src/index.js
@@ -18,9 +18,12 @@ r.add('post', '/validate-email', validateEmail)
 r.add('get', '/validate-ws', validateWS)
 r.add('post', '/', postRoot)
 r.add('post', '/raw', postRoot)
+r.add('get', '/.debug/error', () => {
+  throw new Error('this is the result of handling a request at /.debug/error')
+})
 
 /** @type {import('./bindings.js').ModuleWorker} */
-const worker = {
+export default {
   fetch: async (request, env, ctx) => {
     const context = getContext(request, env, ctx)
     context.log.time('request')
@@ -37,5 +40,3 @@ const worker = {
     }
   },
 }
-
-export default worker