+++ packages/access-client/test/agent.test.js
@@ -1,6 +1,6 @@
 import assert from 'assert'
 import { URI } from '@ucanto/validator'
-import { Agent, connection, createDidMailtoFromEmail } from '../src/agent.js'
+import { Agent, connection } from '../src/agent.js'
 import * as Space from '@web3-storage/capabilities/space'
 import { createServer } from './helpers/utils.js'
 import * as fixtures from './helpers/fixtures.js'
@@ -54,6 +54,46 @@ describe('Agent', function () {
         message: `Agent has no proofs for ${fixtures.alice.did()}.`,
       }
     )
+  })
+
+  it('should allow import a space', async () => {
+    const alice = await Agent.create()
+    const bob = await Agent.create()
+
+    const space = await alice.createSpace('videos')
+    await alice.setCurrentSpace(space.did)
+
+    const proof = await alice.delegate({
+      audience: bob,
+      audienceMeta: { name: 'videos', type: 'app' },
+      abilities: ['*'],
+    })
+
+    await bob.importSpaceFromDelegation(proof)
+    await bob.setCurrentSpace(space.did)
+
+    const proofs = bob.proofs([{ can: 'store/add', with: space.did }])
+    assert(proofs.length)
+  })
+
+  it('should allow import a space with restricted abilities', async () => {
+    const alice = await Agent.create()
+    const bob = await Agent.create()
+
+    const space = await alice.createSpace('videos')
+    await alice.setCurrentSpace(space.did)
+
+    const proof = await alice.delegate({
+      audience: bob,
+      audienceMeta: { name: 'videos', type: 'app' },
+      abilities: ['store/add'],
+    })
+
+    await bob.importSpaceFromDelegation(proof)
+    await bob.setCurrentSpace(space.did)
+
+    const proofs = bob.proofs([{ can: 'store/add', with: space.did }])
+    assert(proofs.length)
   })
 
   it('should invoke and execute', async function () {
@@ -199,7 +239,7 @@ describe('Agent', function () {
       audienceMeta: { name: 'sss', type: 'app' },
     })
 
-    await bob.addProof(delegation)
+    await bob.importSpaceFromDelegation(delegation)
     await bob.setCurrentSpace(space.did)
 
     // should not be able to store/remove - bob only has ability to space/info
@@ -211,13 +251,6 @@ describe('Agent', function () {
           audienceMeta: { name: 'sss', type: 'app' },
         }),
       /cannot delegate capability store\/remove/
-    )
-  })
-
-  it('exports createDidMailtoFromEmail', async () => {
-    assert.deepEqual(
-      createDidMailtoFromEmail('foo@dag.house'),
-      'did:mailto:dag.house:foo'
     )
   })
 })