+++ packages/upload-client/src/store.js
@@ -3,6 +3,24 @@ import * as StoreCapabilities from '@web3-storage/capabilities/store'
 import retry, { AbortError } from 'p-retry'
 import { servicePrincipal, connection } from './service.js'
 import { REQUEST_RETRIES } from './constants.js'
+import fetchPkg from 'ipfs-utils/src/http/fetch.js'
+const { fetch } = fetchPkg
+
+/**
+ *
+ * @param {string} url
+ * @param {import('./types').ProgressFn} handler
+ */
+function createUploadProgressHandler(url, handler) {
+  /**
+   *
+   * @param {import('./types').ProgressStatus} status
+   */
+  function onUploadProgress({ total, loaded, lengthComputable }) {
+    return handler({ total, loaded, lengthComputable, url })
+  }
+  return onUploadProgress
+}
 
 /**
  * Store a DAG encoded as a CAR file. The issuer needs the `store/add`
@@ -67,15 +85,25 @@ export async function add(
     return link
   }
 
+  const fetchWithUploadProgress =
+    /** @type {(url: string, init?: import('./types').FetchOptions) => Promise<Response>} */ (
+      fetch
+    )
+
   const res = await retry(
     async () => {
       try {
-        const res = await fetch(result.url, {
+        const res = await fetchWithUploadProgress(result.url, {
           method: 'PUT',
           mode: 'cors',
           body: car,
           headers: result.headers,
           signal: options.signal,
+          onUploadProgress: options.onUploadProgress
+            ? createUploadProgressHandler(result.url, options.onUploadProgress)
+            : undefined,
+          // @ts-expect-error - this is needed by recent versions of node - see https://github.com/bluesky-social/atproto/pull/470 for more info
+          duplex: 'half',
         })
         if (res.status >= 400 && res.status < 500) {
           throw new AbortError(`upload failed: ${res.status}`)